Make file name consistent with using kebab case
