+++ src/mocks/mock-ims-backend.ts
@@ -7,17 +7,17 @@ import { ModelArchivesPointResponse } from './response/model-archives-point-resp
 import { ModelLink } from './../models/model-link';
 import { ModelArchives } from './../models/model-archives';
 import { QueryBuilderService } from './../providers/query-builder-service';
-import { QueryFragment } from './../models/queryFragment';
+import { QueryFragment } from './../models/query-fragment';
 import { Response, ResponseOptions, RequestMethod } from '@angular/http';
 import { MockBackend } from '@angular/http/testing';
 import { EntryPointResponse } from './response//entry-point-response';
 import { LicensePointResponse } from './response//license-point-response';
-import { LocationResponse } from './response//locationResponse';
-import { TokenResponse } from './response//tokenResponse';
-import { EntriesResponse } from './response/entriesResponse';
-import { ArchiveEntryResponse } from './response/archiveEntryResponse';
-import { ArchiveTableEntry } from '../models/archiveTableEntry';
-import { ArchiveEntry } from '../models/archiveEntry';
+import { LocationResponse } from './response//location-response';
+import { TokenResponse } from './response//token-response';
+import { EntriesResponse } from './response/entries-response';
+import { ArchiveEntryResponse } from './response/archive-entry-response';
+import { ArchiveTableEntry } from '../models/archive-table-entry';
+import { ArchiveEntry } from '../models/archive-entry';
 import { Token } from '../models/token';
 import { Link } from '../models/link';
 import { Credential } from '../models/credential';
