+++ src/models/archive-table-entry.ts
