+++ src/models/query-fragment.ts
