+++ src/providers/ims-service.ts
@@ -3,13 +3,13 @@ import { Http, Response } from '@angular/http';
 import { Observable } from 'rxjs/Observable';
 import { Info } from '../models/info';
 import { Credential } from './../models/credential';
-import { ImsHeaders } from '../models/imsHeaders';
+import { ImsHeaders } from '../models/ims-headers';
 import { EntryPoint } from '../models/entry-point';
 import { LicensePoint } from '../models/license-point';
 import { EntriesPoint } from '../models/entries-point';
 import { Token } from '../models/token';
-import { ArchiveEntry } from '../models/archiveEntry';
-import { ArchiveTableEntry } from '../models/archiveTableEntry';
+import { ArchiveEntry } from '../models/archive-entry';
+import { ArchiveTableEntry } from '../models/archive-table-entry';
 import { Link } from '../models/link';
 import { Filter } from '../models/filter';
 import { ModelArchives } from './../models/model-archives';
