+++ src/providers/query-builder-service.spec.ts
@@ -1,4 +1,4 @@
-import { QueryFragment } from './../models/queryFragment';
+import { QueryFragment } from './../models/query-fragment';
 import { QueryBuilderService } from './query-builder-service';
 import { TestBed, inject, async } from '@angular/core/testing';
