+++ src/providers/upload-service.ts
@@ -6,7 +6,7 @@ import 'rxjs/add/observable/fromPromise';
 import { Credential } from '../models/credential';
 import { Entry } from '../models/entry';
 import { Image } from '../models/image';
-import { ImsHeaders } from '../models/imsHeaders';
+import { ImsHeaders } from '../models/ims-headers';
 import { Token } from '../models/token';
 import { Observable } from 'rxjs/Observable';
 import { TokenService } from './token-service';
